requirements.txt
kafka-python
google-cloud-bigquery